Correctly handle self-closing SVG tags
SVG and MathML in HTML doesn't trigger a full HTML parser,
but it does force self-closing tags to be recognized.

In contrast, the old approach of treating SVG shapes
as void elements resulted in serious errors.
See issue #187

I'm running tests from cheerio/dom-serializer,
but if they could be adapted to this repo,
that would be even better.