fix(parser): Fix `endIndex` of text events being off by 1
While other events use the end index of a particular sequence (eg. a `>` in a tag) to signify the end index, text events use the start of something else as an index. This led to an off-by-one error in the previous release, where text events would have an index that was too high by 1. This fixes this.