tests: update Events/07 test to clarify interpretation of tag end slashes
Enhancements to the `Events/07-self-closing.json` test to better cover and
demonstrate the interpretation of tag end slashes (e.g. `/>`)

`07-self-closing.json` is replaced with a set of tests 07a to 07h where
the input and context are identical except for one isolated change. The
outputs can thereby be diffed, showing the effect of the isolated change.

Adds test coverage for the `recognizeSelfClosing` option.